LayoutTests/requestidlecallback/requestidlecallback-deadline-cap-by-rendering-update-expected.txt

This tests that the deadline can be reached.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS idleDeadline.timeRemaining() <= preferredRenderingUpdateInterval is true
PASS didRunIdleCallback became true
PASS successfullyParsed is true

TEST COMPLETE
